+++ common/PEQvelocities.c
@@ -137,7 +137,7 @@ void InitPEQvelocities(
 static void scheduled_change_velocity(int64_t time, const union SuperType *args){
   const struct Tracks *track = args[0].const_pointer;
   const struct Notes  *note  = args[1].const_pointer;
-  int            x     = args[2].int_num;
+  int            x           = (int)args[2].int_num;
 
   //printf("   Change velocity to %d\n",x);
